pyproject.toml
[tool.poetry]
name = "bot"
version = "1.0a"
description = "The bot service for poker."
authors = ["likecompany"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
likeinterface = "^2.5"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
aiogram = {extras = ["redis"], version = "^3.1.1"}
apscheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
black = "^23.3.0"
ruff = "^0.0.287"
isort = "^5.12.0"
mypy = "^1.2.0"

[tool.ruff]
line-length = 99
select = [
    "C4",
    "E",
    "F",
    "T10",
    "T20",
    "Q",
    "RET",
]
ignore = ["E501"]
exclude = [".git", ".venv"]
force-exclude = true

[tool.black]
line-length = 99
target-version = ["py38", "py39", "py310", "py311"]
exclude = '''
(
    \.git
    \.ruff_cache
    \.venv
    \.mypy_cache
)
'''

[tool.isort]
profile = "black"
src_paths = ["app"]
skip = ["alembic"]

[tool.mypy]
plugins = ["pydantic.mypy"]
disable_error_code = ["import"]
show_error_codes = true
show_error_context = true
namespace_packages = true
